requirements.txt
falcon>=1.4.1
numpy
scipy
tensorflow
pyyaml
argparse
symspellpy
